changed how new bandwidths are manipulated
The new bandwidths are now altered by a separate callback. This callback uses the information from the figure object to identify if a new bandwidth was selected by the user. It looks for a change in the range slider. Then uses that to create a new vertical line or modify an existing one.